change input prop of H5WasmProvider to take File or URL
The single input property of H5WasmProvider is altered to be either a File object (result of choosing in HTML file input element) or a URL representing a
remote file to be loaded over http(s).  Some cleanup is required if the H5WasmProvider is re-rendered, and this is done in the useEffect callback.