Invoke `SelectionTool` render prop/cbs with transformed selection
Passing the effective selection in addition to the  raw selection allows the computation of the zoom selection in `SelectToZoom` to be done in a `transformSelection` callback.

The transformed/effective selection is a derived state of the raw selection state. Previous attempts had the effective selection stored inside the same state as the raw selection, but this is not needed.